requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo
pymongo[srv]==3.6

# add at the end when project is done
# it will create all the packages
##-e .
